requirements.txt
streamlit
groq
replicate
python-dotenv
pillow
numpy
